Add tenant_id to queries using rewrite instead of default_scope/unscoped
Previously we relied on default_scope and a monkey patch to automatically
add the tenant_id to queries. This was brittle in some cases, in
particular when used with table aliases in complex JOINs.

Instead, walk the Arel tree using a visitor and add tenant_id for the
relations that need it.

In addition this changes the existing before_validation(on: :create) into
an after_initialize, to match the initialization behaviour previously
provided by the default_scope.

Author: Ben Olive (https://github.com/sionide21)