Fixes automatic inverse of on singular associations
This changes the call to `belongs_to` to pull the `:foreign_key` option
from the options, instead of defaulting to the foreign_key generated by
the call to `MultiTenant.partition_key(tenant_name)` on line 39. We do
this because passing a foreign key option will disable the automatic
inverse of reflection within ActiveRecord. See the links below for the
lines in ActiveRecord that prevent the automatic inverse calculation
with this option.

https://github.com/rails/rails/blob/master/activerecord/lib/active_record/reflection.rb#L595

and

https://github.com/rails/rails/blob/master/activerecord/lib/active_record/reflection.rb#L671-L676