Add test for `find`
Attempting to `find` a record for a tenant other than the current tenant with raise an `ActiveRecord::NotFound` error.

I am not sure how to test code that behaves differently with `MultiTenant.enable_write_only_mode` and a quick search didn't turn up any examples. Maybe that feature is untested?