Fix find and find_by caching issues
This builds on work by multiple contributors, and fixes issues where the
tenant_id would be cached across different tenant contexts for find
and find_by methods. This issue was only present with prepared
statements turned on.

Note that the mechanism to solve this is slightly different for Rails 4
and 5:

* Rails 4: Disable any caching for find and find_by methods
* Rails 5: Explicitly add the current_tenant_id into the cache key

This also ensures that we test both prepared statements on and off
on Travis.