Add client and server sidekiq middleware to sidekiq middleware chain
Today we don't have any public documentation for the gem on how/where
to add the Sidekiq middlewares to ensure that `multi_tenant` object
is attached to the Sidekiq job when enqueuing. It can take bit of
figuring out to do it the right away.

Instead, with this PR, the experience can be a bit nicer because
as a consumer of this gem we won't have to think about how/where to add
this.

All we will need is a require `'activerecord-multi-tenant/sidekiq'`

Without the addition of middleware, the `multi_tenant` object doesn't get attached
to the job