Only attempt to reload with MultiTenant when parition_key is present
When using gems like graphql or similar, they ca returns records from association
which do not contain the `partition_key` column. In such cases, we'd see `ActiveModel::MissingAttributeError: missing attribute` error raised.

This change makes it so that we just call the upstream/original method (`original_method_name`)
when there is no `partition_key` present in the record. Thus, resulting in no exception.

Also, adding a proof guard to ensure `partition_key` function is present on the class (`owner.class`).
That is also another possibility of errors, especially when mocking models or overriding models
in specs. Due to the way the ActiveRecord is overriden, we can also run into errors is "NoMethodError"
for `partition_key`